fix arm64 page_present type issue

+++ arm64.c
@@ -3087,7 +3087,7 @@ arm64_translate_pte(ulong pte, void *physaddr, ulonglong unused)
 	char ptebuf[BUFSIZE];
 	char physbuf[BUFSIZE];
         char *arglist[MAXARGS];
-	int page_present;
+	ulong page_present;
 
 	paddr = pte & PHYS_MASK & (s32)machdep->pagemask;
        	page_present = pte & (PTE_VALID | machdep->machspec->PTE_PROT_NONE);
